requirements.txt
asttokens==3.0.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
decorator==5.2.1
executing==2.2.0
Flask==3.1.0
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jsonpickle==4.0.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
networkx==3.4.2
numpy==2.2.3
parso==0.8.4
prompt_toolkit==3.0.50
pure_eval==0.2.3
Pygments==2.19.1
pyvis==0.3.2
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.1.3
